requirements.txt
# Python ML Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
PyYAML>=6.0
albumentations>=1.3.0
pycocotools>=2.0.6

# ONNX & Runtime
onnx>=1.14.0
onnxruntime-gpu>=1.15.0

# Python bindings for C++ (if needed)
pybind11>=2.11.0

# Utilities
tqdm>=4.65.0
tensorboard>=2.13.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0
scikit-learn>=1.3.0
mlflow>=2.8.0
requests>=2.31.0

# Note: C++ dependencies (TensorRT, CUDA, CMake) must be installed separately
# See README.md Prerequisites section for installation instructions
